v1.0.2 - DM support
- Quickly changed the way that loading commands get logged.
- Changed commands for DM support.
- On the way to adding DM support I forgot to add the testing bot to the emojis servers and spent 3 hours typing in a workaround only to realise i am a big dumb idiot.

+++ commands/level.js
@@ -5,6 +5,7 @@ const gdtools = require('../functions/gdFunctions.js');
 
 module.exports.run = async (bot, message, args) => {
 	if (!/[0-9]/.test(args[1])) {
+		if (!message.guild) return message.channel.send(`This sub-command may only be used in servers, due to Discord limitations.`);
 		message.delete();
 		const e = config.emojis;
 		let search = await tools.getArgs(args, 1);
